Added static build scripts for ios/android/mac
Fixes:
- Common: fixed utf encoding for 3-bytes chars
- Common: NetworkHandle: fixed missed cleanup
- Serenity: AllocStack: fixed server getter (inverted priority)
- Mac: implemented makefile builder

+++ stappler/src/features/rich_text/renderer/SPRichTextResult.h
@@ -40,7 +40,7 @@ public:
 	};
 
 	struct BoundIndex {
-		size_t idx = maxOf<int64_t>();
+		size_t idx = maxOf<size_t>();
 		float start;
 		float end;
 		int64_t page;
